requirements.txt
# Core inference dependencies
numpy==1.26.4
torch==2.2.2
torchvision==0.17.2
opencv_python==4.9.0.80
Pillow==10.3.0
scipy==1.13.1

# YOLO detection
ultralytics==8.1.24
einops==0.8.0

# Text and image processing
paddleocr==2.7.5
paddlepaddle_gpu==2.6.1.post120

# Utilities
tqdm==4.66.2
pandas==2.2.2

# Screen capture (for MuMu adapter)
mss==9.0.1
pyautogui==0.9.53

# Optional visualization
matplotlib==3.8.2
